all done

+++ src/app/pages/invoice/invoice.component.ts
@@ -47,6 +47,7 @@ export class InvoiceComponent implements OnInit {
   paymentGate:boolean=false;
   paymentSessionId="session_865qQHlSgTzVXcRTesyDrYLn2e65s0u5bunLqby5PN9UAa_VIahEU2etSR_oqZLV1sWHV3nFoNcl";
   @ViewChild('invoiceId') invoiceId: ElementRef;
+  showHeader: boolean;
   constructor(private messageService: MessageService,public paymentService:PaymentService,
     public confirmationService: ConfirmationService,public invoiceitemser:InvoiceItemService,public fb:FormBuilder,public invoiceService:InvoiceService,public customerser:CustomerService) { 
   }
@@ -63,20 +64,20 @@ export class InvoiceComponent implements OnInit {
   }
   createInvoiceFormgroup(){
     this.invoiceGroup=this.fb.group({
-      _id:this.fb.control(''),
+    
+      id:this.fb.control(''),
       paymentType:this.fb.control('',Validators.required),
-      createdOn:this.fb.control('',Validators.required),
-     invoicedate:this.fb.control('',Validators.required),
-     totalamount:this.fb.control('',Validators.required),
+      createdAt:this.fb.control('',Validators.required),
+      invoicedate:this.fb.control('',Validators.required),
+      totalamount:this.fb.control('',Validators.required),
      invoiceno:this.fb.control('',Validators.required),
-      customerId:this.fb.control('',Validators.required),
-    link_currency:this.fb.control({value:'', disabled: true}),
+      customer_id:this.fb.control('',Validators.required),
+    link_currency:this.fb.control({value:'INR', disabled: true}),
       link_purpose:this.fb.control({value:'', disabled: true}),
       link_notify:this.fb.group({
         send_sms:this.fb.control({value:'', disabled: true}),
         send_email:this.fb.control({value:'', disabled: true})
       }),
-      methodtype:this.fb.control(''),
       link_partial_payments:this.fb.control({value:'', disabled: true}), 
       link_minimum_partial_amount:this.fb.control({value:'', disabled: true}),      
       // customer_details:this.fb.group({
@@ -85,44 +86,38 @@ export class InvoiceComponent implements OnInit {
       //   customer_phone:this.fb.control(''),
       //   customer_email:this.fb.control('')
       // }),
-      __v:this.fb.control(''),
-      custdata:this.fb.group({
-       name:this.fb.control(''),
-       email:this.fb.control(''),
-       contact:this.fb.control(''),
-       address:this.fb.group({
-         line1:this.fb.control(''),
-         line2:this.fb.control(''),
-         pincode:this.fb.control(''),
-         city:this.fb.control(''),
-         state:this.fb.control(''),
-         country:this.fb.control(''),
-       })
-      }),
+      // custdata:this.fb.group({
+      //  name:this.fb.control(''),
+      //  email:this.fb.control(''),
+      //  contact:this.fb.control(''),
+      //  address:this.fb.group({
+      //    line1:this.fb.control(''),
+      //    line2:this.fb.control(''),
+      //    pincode:this.fb.control(''),
+      //    city:this.fb.control(''),
+      //    state:this.fb.control(''),
+      //    country:this.fb.control(''),
+      //  })
+      // }),
       item:this.fb.array([
-     
       ]),
+      methodtype:this.fb.control(''),
  })
  this.invoiceGroup.valueChanges.subscribe((x)=>{
   console.log(x.link_partial_payments);
   if(x && x.paymentType){
-
-    // x.link_currency.value= 'INR',
     this.invoiceGroup.get('link_currency').enable({onlySelf:true});
     this.invoiceGroup.get('link_purpose').enable({onlySelf:true});
     this.invoiceGroup.get('link_partial_payments').enable({onlySelf:true});
     this.invoiceGroup.get('link_notify.send_sms').enable({onlySelf:true});
     this.invoiceGroup.get('link_notify.send_email').enable({onlySelf:true});
-    
     if(x && x.link_partial_payments == true){
       this.invoiceGroup.get('link_minimum_partial_amount').enable({onlySelf:true});      
     }
     else{
       this.invoiceGroup.get('link_minimum_partial_amount').disable({onlySelf:true});
     }  
-  
   }
-
 })
   }
   get itemInvoice() {
@@ -130,34 +125,27 @@ export class InvoiceComponent implements OnInit {
   }
 
   editInvoice(product){
+    this.showHeader=true;
     this.createInvoiceFormgroup();
     this.invoiceDialog=true;
-    console.log(product,'edit');
-    // product.item.forEach((element,value) => {
-    //   this.itemInvoice.push(this.fb.control(element))
-    //   let d={data:element,val:value};
-    //   console.log(d);
-    // });
-    
+    console.log(product,'edit',product.createdAt);
     this.invoiceGroup.patchValue({
-      _id:product._id,
-      __v:product.__v,
+      id:product.id,
       paymentType:product.paymentType,
-      link_currency:product.paymentDetail.link_currency,
-      link_purpose:product.paymentDetail.link_purpose,
-
+      link_currency:product.payDetail.link_currency,
+      link_purpose:product.payDetail.link_purpose,
       totalamount:product.totalamount,
-      createdOn:product.createdOn,
+      // createdOn:product.createdAt,
+      createdAt:product.createdAt,
       invoicedate:new Date(product.invoicedate),
       invoiceno:product.invoiceno,
-      customerId:this.mapCustomerwithid(product.customerId),
+      customer_id:this.mapCustomerwithid(product.customer_id),
    
     })
     this.invoiceitem.clear();
 
-    for (let i of product.item) {
+    for (let i of product.invoiceItems) {
       if(i){
-        
         this.invoiceitem.push(this.createItem(i));
       } 
       
@@ -165,31 +153,31 @@ export class InvoiceComponent implements OnInit {
     // this.invoiceGroup.patchValue({item:product.item},)
     console.log(this.invoiceGroup.value,'patched');
   } 
+
+
     getReadPdf(obj){
       this.invoiceService.getInvoiceReadablePDF(obj.invoiceno).subscribe((x:any)=>{
         const url = window.URL.createObjectURL(x); // create a URL object from the blob
         const link = document.createElement('a'); // create an anchor tag
         link.href = url; // set its href attribute to the blob's URL
-        link.download = 'invoice.pdf'; // set its download attribute to the file name
+        link.download = `invoice_${obj.invoiceno}.pdf`; // set its download attribute to the file name
         link.click();
       }),(err:any)=>{
         this.messageService.add({severity:'error', summary: 'Failed', detail: 'Failed to Download', life: 3000});
       } 
     }
+
+    
   downloadPdf(obj){ 
     let PDF = new jsPDF('p', 'mm', 'a4');
     this.invoicepdf=true;
     this.invoiceService.getInvoicePdfbyNo(obj.invoiceno).subscribe((x:any)=>{
-      
       this.invoicecontent = x;
-      
       this.invoiceId.nativeElement.innerHTML =this.invoicecontent;
       let DATA: any = document.getElementById('invoiceId');
         // PDF.html(DATA).then(x=>{PDF.save('invoice-generated.pdf')})
-   
       html2canvas(DATA).then((canvas) => {
         console.log(canvas);
-        
         let fileWidth = 208;
         let fileHeight = (canvas.height * fileWidth) / canvas.width;
         const FILEURI = canvas.toDataURL('image/png');
@@ -210,11 +198,16 @@ export class InvoiceComponent implements OnInit {
     });
     })    
   }
+
+
+
   filterCountry(event) {
     let filtered : any[] = [];
     let query = event.query;
     for(let i = 0; i < this.custArray.length; i++) {
         let cdata = this.custArray[i];
+        console.log(cdata,'cdata check');
+        
         if (cdata.name.toLowerCase().indexOf(query.toLowerCase()) == 0) {
           // console.log(cdata,'customer data');
           filtered.push(cdata);
@@ -222,36 +215,40 @@ export class InvoiceComponent implements OnInit {
     }
     this.filteredData = filtered;
 }
-  selectRow(checkValue) {
-    if (checkValue) {
-      this.deleteId = this.invoiceArr.map(value => value._id);
-    } else {
-      this.deleteId = [];
-    }
-  }
+
   deleteSelectedProducts(){
-    console.log(this.deleteId);
-    let id =this.deleteId.map(x=>x._id);
-    console.log(id);
-    
-    this.invoiceService.deleteSelectedInvoice(id).subscribe((x:any)=>{
-      if(x.succes){
-        this.messageService.add({severity:'error', summary: 'Deleted Invoice', detail: x.deleted, life: 3000});
-         this.deleteId=null
-         id=[] 
-        this.getinvoice();
+  this.confirmationService.confirm({
+      message: 'Are you sure you want to delete select Invoice?',
+      header: 'Confirm',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        console.log(this.deleteId);
+        let id =this.deleteId.map(x=>x.id);
+        console.log(id); 
+        this.invoiceService.deleteSelectedInvoice(id).subscribe((x:any)=>{
+          if(x.success){
+            this.messageService.add({severity:'error', summary: 'Deleted Invoice', detail: x.deleted, life: 3000});
+             this.deleteId=null
+             id=[] 
+            this.getinvoice();
+          }
+        })
       }
-    })
+  });
+    
+   
+  
   }
+
+  
  deleteInvoice(product){
-    // this.invoiceService.deleteInvoice()
     this.confirmationService.confirm({
       message: 'Are you sure you want to delete the this Invoice?',
       header: 'Confirm',
       icon: 'pi pi-exclamation-triangle',
       accept: () => {
-        this.invoiceService.deleteInvoice(product._id).subscribe((x:any)=>{
-          if(x.succes){
+        this.invoiceService.deleteInvoice(product.id).subscribe((x:any)=>{
+          if(x.success){
             this.messageService.add({severity:'error', summary: 'Deleted', detail: 'Invoice Deleted', life: 3000});
             this.getinvoice();
           }
@@ -266,9 +263,11 @@ getinvoice(){
     this.loading = false; 
   })
 }
+
 get invoiceitem() {
 	return this.invoiceGroup.get('item') as FormArray;
 }
+
 
 createPayementGateway(sessioId,component,parent){
   this.getinvoice();
@@ -281,17 +280,27 @@ createPayementGateway(sessioId,component,parent){
   }
   let dropinConfig = {
     components:component,
-    onSuccess: function(data){
+    onSuccess: (data)=>{
         parent.innerHTML='';
       if (data.order && data.order.status == "PAID") {
         console.log(data);
-        
-        alert(JSON.stringify(data));
+        let d = {
+          status:data.order.status,
+          amount:parseInt(data.transaction.transactionAmount),
+          transactionId:data.transaction.transactionId
+      }
+      parent.innerHTML='';    
+      this.invoiceService.updateStatusAfterPayment(data.order.orderId,d).subscribe((x:any)=>{
+        this.messageService.add({severity:'success', summary: 'Success', detail: 'Payment Recieved', life: 3000});
+        this.getinvoice();
+      },(err)=>{
+        this.messageService.add({severity:'error', summary: 'Failed', detail: 'Payment Failed', life: 3000});
+      })
       }
     },
     onFailure: function(data){
       console.log(data);
-      
+      parent.innerHTML='';
         alert(data.order.errorText);
     },
     style: {
@@ -311,20 +320,22 @@ cashfree.drop(parent,dropinConfig);
 addItem(){
   this.invoiceitem.push(this.createItem());
 }
+
   getCustomer(){
     this.customerser.getCustomer().subscribe((x)=>{
       let data= x.data    
       this.custArray = data.map((x:any)=>{
-        const d ={name:x.name,id:x._id} 
+        const d ={name:x.firstName,id:x.id} 
         return d;
       })
-      console.log(this.custArray);
+      console.log(this.custArray,'cust data check');
     })
   }
+
   getInvoiceItem(){
     this.invoiceitemser.getInvoiceItem().subscribe((x:any)=>{
         this.itemArr= x.data.map((x:any)=>{
-          const d = {name:x.name,id:x._id}
+          const d = {name:x.itemname,id:x.id}
           return d;
         })
         console.log(this.itemArr,'item array');
@@ -333,31 +344,33 @@ addItem(){
   }
   createItem(data?:any): FormGroup {
     return this.fb.group({
-      invoice_itemId:this.fb.control(data && data.invoice_itemId ? data.invoice_itemId : '' ,Validators.required),
-      invoice_data:this.fb.control(''),
+      item_id:this.fb.control(data && data.item_id ? data.item_id : '' ,Validators.required),
+      // invoice_data:this.fb.control(''),
       quantity:this.fb.control(data && data.quantity ? data.quantity : '',Validators.required),
     });
   }
+
   openNew(){
-    // this.createInvoiceFormgroup()
     this.invoiceitem.clear()
     this.invoiceGroup.reset({});
     this.invoiceObj={}
     this.invoiceDialog=true;
     this.invoiceitem.push(this.createItem())
   }
+
   getPdf(obj){
     this.invoicepdf=true;
     this.invoiceService.getInvoicePdfbyNo(obj.invoiceno).subscribe((x:any)=>{
       this.invoicecontent = x;
+
       this.invoiceId.nativeElement.innerHTML =this.invoicecontent;
       
     })
   }
+
+
   mapCustomerwithid(ids){
     const d= this.custArray.filter((x,value)=>{
-      // console.log(x,value);
-      
       if(x.id == ids){
         return value
 
@@ -365,11 +378,12 @@ addItem(){
     })  
     return d[0];
   }
-  getPaymentStatus(obj,invoice):any{
-    this.paymentService.getPaymentStatus(obj).subscribe((x:any)=>{
+
+  getPaymentStatus(id):any{
+    this.paymentService.getPaymentStatus(id).subscribe((x:any)=>{
       if(x.success){
         console.log(x.data.link_status);
-        let id = invoice.paymentDetail._id;
+        // let id = invoice.payDetail.id;
         let data = x.data
         this.paymentService.updateStatus(id,data).subscribe((x:any)=>{
           this.messageService.add({severity:'warn', summary: 'Successful', detail: 'Status Updated', life: 3000});
@@ -384,22 +398,28 @@ addItem(){
     })
 
   }
+
+
   fnSubmit(){
+    this.showHeader=false;
     this.submitted=true;
-    const itemarray = this.invoiceGroup.get('item').value
+    let no;
+    const itemarray = this.invoiceGroup.get('item').value;
     const items = itemarray.map((x)=>{
-      const d = {invoice_itemId:x.invoice_itemId,quantity:x.quantity}
+      const d = {item_id:x.item_id,quantity:x.quantity}
       return d;
     })
-    const d=this.invoiceGroup.get('customerId').value
+    const d=this.invoiceGroup.get('customer_id').value
     console.log(d);
+
     if(d != undefined){
       this.invoiceGroup.patchValue({
-        customerId:d.id
+        customer_id:d.id
       })
     }
+
     let invoiceData = {
-      customerId:this.invoiceGroup.get('customerId').value,
+      cust_id:this.invoiceGroup.get('customer_id').value,
       item:items,
       invoicedate:this.invoiceGroup.get('invoicedate').value,
       invoiceno:this.invoiceGroup.get('invoiceno').value,
@@ -412,32 +432,31 @@ addItem(){
       link_minimum_partial_amount:this.invoiceGroup.get('link_minimum_partial_amount').value,
       link_currency:this.invoiceGroup.get('link_currency').value.toUpperCase(),
       link_purpose:this.invoiceGroup.get('link_purpose').value,
+      methodtype:this.invoiceGroup.get('methodtype').value,
     }
-    let datacheck = this.invoiceGroup.get('__v').value
-    let id = this.invoiceGroup.get('_id').value; 
-    console.log(this.invoiceGroup.value,Array.isArray(itemarray),itemarray,id);
-    if(this.invoiceGroup.valid && (this.submitted && (id !== '' || datacheck))){
-     this.invoiceGroup.removeControl('_id');
+    let cust_id = this.invoiceGroup.get('customer_id').value;
+    let invoiceno = this.invoiceGroup.get('id').value;
+    console.log(this.invoiceGroup.value,Array.isArray(itemarray),itemarray,invoiceno,'All value check');
+
+    if(this.invoiceGroup.valid && (this.submitted && (cust_id !== '' || (cust_id && invoiceno)))){
+     this.invoiceGroup.removeControl('id');
      this.invoiceGroup.removeControl('custdata');
      this.invoiceGroup.removeControl('totalamount')
       console.log(this.invoiceGroup.value);
       
-    this.invoiceService.updateInvoice(id,this.invoiceGroup.value).subscribe((x)=>{
-      if(x.succes){
+    this.invoiceService.updateInvoice(invoiceno,this.invoiceGroup.value).subscribe((x)=>{
+      if(x.success){
         this.messageService.add({severity:'custom', summary: 'Successful', detail: 'Invoice Updated', life: 3000});
         this.getinvoice();
       }
     })
     }
     else{
-
       this.invoiceService.postInvoice(invoiceData).subscribe((x:any)=>{
-        console.log(x,'data');
-
-        if(x.succes){
-          if(x.paymentType==2){
-            this.messageService.add({severity:'success', summary: 'Successful', detail: 'Invoice Created', life: 3000}); 
-            let sesID = x.data.payment_session_id;
+        if(x.success){
+          if(x.paymentType == 2){
+            this.messageService.add({severity:'success', summary: 'Successful', detail: 'Invoice Created with gateway', life: 3000}); 
+            let sesID = x.payment_session_id;
             let component = this.invoiceGroup.get('methodtype').value;
             let parent = document.getElementById("drop_in_container");
             this.createPayementGateway(sesID,component,parent)
